Package restructuration , adding @Incubation annotation on classes, avoiding create mainDistZip task.
