+++ subprojects/plugins/src/integTest/groovy/org/gradle/api/plugins/DistributionPluginIntegrationTest.groovy
@@ -27,6 +27,10 @@ class DistributionPluginIntegrationTest extends WellBehavedPluginTest {
         "distribution"
     }
 
+    String getMainTask() {
+        return "distZip"
+    }
+
     def createTaskForCustomDistribution() {
         when:
         buildFile << """
