+++ subprojects/plugins/src/main/groovy/org/gradle/api/distribution/internal/DefaultDistribution.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2010 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.gradle.api.distribution.internal;
+
+import org.gradle.api.distribution.Distribution;
+
+/**
+ * Allow user to declare a distribution.
+ * @author scogneau
+ */
+public class DefaultDistribution implements Distribution {
+    public DefaultDistribution(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    private String name;
+}
