+++ subprojects/plugins/src/main/resources/META-INF/gradle-plugins/distribution.properties
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-implementation-class=org.gradle.api.plugins.DistributionPlugin
+implementation-class=org.gradle.api.distribution.plugins.DistributionPlugin
